backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-dotenv
langchain
wcwidth
scikit-learn
pandas
numpy
matplotlib
seaborn

